order_service/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["ChWasam <wasam.developer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.18"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
psycopg = {extras = ["binary"], version = "^3.1.19"}
protobuf = "^5.27.0"
aiokafka = "^0.10.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
